+++ R/policy.R
@@ -24,57 +24,23 @@
 #' # policy with value function, optimal action and transitions for observations.
 #' policy(sol)
 #' @export
-policy <- function(x) {
-  if (inherits(x, "MDP"))
-    .policy_MDP(x)
-  else
-    .policy_POMDP(x)
-}
+policy <- function(x) x$solution$policy
 
-.policy_POMDP <- function(x) {
-  .solved_POMDP(x)
-  
+.policy_MDP_from_POMDP <- function(x) {
   pg <- x$solution$pg
-  alpha <- x$solution$alpha
+  bs <- x$model$observation_prob[['*']]
   
-  for (i in 1:length(pg)) {
-    pg[[i]] <- cbind(alpha[[i]],
-      pg[[i]][, "action", drop = FALSE])
-  }
-  
-  pg
-}
-
-# FIXME: This is a MDP policy
-
-# create belief states to MDP (identity matrix)
-.MDP_belief_states <- function(model, projection = NULL) {
-  belief_states <-
-    diag(1,
-      nrow = length(model$model$states),
-      ncol = length(model$model$states))
-  colnames(belief_states) <- model$model$states
-  if (!is.null(projection)) {
-    # FIXME: convert state names!
-    belief_states <- belief_states[projection,]
-  }
-  
-  belief_states
-}
-
-.policy_MDP <- function(x) {
-  .solved_POMDP(x)
-  
-  pg <- x$solution$pg
-  bs <- .MDP_belief_states(x)
-  
-  for (i in 1:length(pg)) {
-    rew <- reward(x, belief = bs, epoch = i)
-    pg[[i]] <- pg[[i]][rew$pg_node,]
-    pg[[i]] <-
-      cbind(data.frame(state = colnames(bs), reward = rew$reward),
-        pg[[i]][,-1])
-  }
-  
-  pg
+  # create a list ith epochs
+  lapply(
+    seq_along(pg),
+    FUN = function(i) {
+      rew <- reward(x, belief = bs, epoch = i)
+      data.frame(
+        state = colnames(bs),
+        U = rew$reward,
+        action = rew$action,
+        pg_node = rew$pg_node
+      )
+    }
+  )
 }