+++ tests/testthat/test-solve_SARSOP.R
@@ -26,5 +26,4 @@ if(sarsop::assert_has_appl()) {
   
   # check alpha vectors
   expect_true(all((pol_SARSOP[,1:2] - pol_POMDP[,1:2]) < 1e-3))
-
 }